feat(routing): full protobuf Error decode for downstream classification
Replace placeholder protobuf_first_int + error_code: 0 in the Error
branch with a full prost decode of crate::proto::ErrorMessage. Both
request_id and error_code now reflect real fields, enabling warning
classification on the protobuf path (PR 1 of warning-message-routing).

Extract routing.rs inline tests to sibling routing_tests.rs to match
the rest of transport/ (rule 14).
