+++ src/transport/routing_tests.rs
@@ -1,0 +1,158 @@
+use super::*;
+use crate::messages::ResponseMessage;
+
+#[test]
+fn test_determine_routing_by_request_id() {
+    // Create a mock message with request ID (AccountSummary = 63)
+    let message_str = "63\01\0123\0DU123456\0AccountType\0ADVISOR\0USD\0";
+    let message = ResponseMessage::from(message_str);
+
+    match determine_routing(&message) {
+        RoutingDecision::ByRequestId(id) => assert_eq!(id, 123),
+        routing => panic!("Expected ByRequestId routing, got {routing:?}"),
+    }
+}
+
+#[test]
+fn test_determine_routing_error_old_format() {
+    // Old format (server_version < ERROR_TIME): message_type|version|request_id|error_code|error_msg
+    let message_str = "4\02\0123\0200\0No security definition found\0";
+    let message = ResponseMessage::from(message_str);
+
+    match determine_routing(&message) {
+        RoutingDecision::Error { request_id, error_code } => {
+            assert_eq!(request_id, 123);
+            assert_eq!(error_code, 200);
+        }
+        routing => panic!("Expected Error routing, got {routing:?}"),
+    }
+}
+
+#[test]
+fn test_determine_routing_error_new_format() {
+    // New format (server_version >= ERROR_TIME): message_type|request_id|error_code|error_msg
+    let message_str = "4\0123\0200\0No security definition found\0\0";
+    let message = ResponseMessage::from(message_str).with_server_version(crate::server_versions::ERROR_TIME);
+
+    match determine_routing(&message) {
+        RoutingDecision::Error { request_id, error_code } => {
+            assert_eq!(request_id, 123);
+            assert_eq!(error_code, 200);
+        }
+        routing => panic!("Expected Error routing, got {routing:?}"),
+    }
+}
+
+#[test]
+fn test_determine_routing_error_protobuf() {
+    // Protobuf Error with id=42 and error_code=2100 (warning) — full decode must populate both fields.
+    let envelope = crate::proto::ErrorMessage {
+        id: Some(42),
+        error_time: None,
+        error_code: Some(2100),
+        error_msg: Some("Market data farm connection is OK".to_string()),
+        advanced_order_reject_json: None,
+    };
+    let mut raw_bytes = Vec::new();
+    prost::Message::encode(&envelope, &mut raw_bytes).expect("encode error envelope");
+
+    let message = ResponseMessage::from_protobuf(IncomingMessages::Error as i32, raw_bytes, crate::server_versions::PROTOBUF);
+
+    match determine_routing(&message) {
+        RoutingDecision::Error { request_id, error_code } => {
+            assert_eq!(request_id, 42);
+            assert_eq!(error_code, 2100);
+        }
+        routing => panic!("Expected Error routing, got {routing:?}"),
+    }
+}
+
+#[test]
+fn test_determine_routing_error_protobuf_unspecified_id() {
+    // Protobuf Error with no id (global notice) decodes to UNSPECIFIED_REQUEST_ID.
+    let envelope = crate::proto::ErrorMessage {
+        id: None,
+        error_time: None,
+        error_code: Some(2104),
+        error_msg: Some("Market data farm connection is OK".to_string()),
+        advanced_order_reject_json: None,
+    };
+    let mut raw_bytes = Vec::new();
+    prost::Message::encode(&envelope, &mut raw_bytes).expect("encode error envelope");
+
+    let message = ResponseMessage::from_protobuf(IncomingMessages::Error as i32, raw_bytes, crate::server_versions::PROTOBUF);
+
+    match determine_routing(&message) {
+        RoutingDecision::Error { request_id, error_code } => {
+            assert_eq!(request_id, UNSPECIFIED_REQUEST_ID);
+            assert_eq!(error_code, 2104);
+        }
+        routing => panic!("Expected Error routing, got {routing:?}"),
+    }
+}
+
+#[test]
+fn test_determine_routing_shared_message() {
+    // ManagedAccounts message (type 15)
+    let message_str = "15\01\0DU123456,DU234567\0";
+    let message = ResponseMessage::from(message_str);
+
+    match determine_routing(&message) {
+        RoutingDecision::SharedMessage(msg_type) => {
+            assert_eq!(msg_type, IncomingMessages::ManagedAccounts);
+        }
+        routing => panic!("Expected SharedMessage routing, got {routing:?}"),
+    }
+}
+
+#[test]
+fn test_is_warning_error() {
+    // Test range boundaries
+    assert!(is_warning_error(2100));
+    assert!(is_warning_error(2169));
+
+    // Test some values in the middle
+    assert!(is_warning_error(2119));
+    assert!(is_warning_error(2150));
+
+    // Test values outside the range
+    assert!(!is_warning_error(2099));
+    assert!(!is_warning_error(2170));
+    assert!(!is_warning_error(200));
+    assert!(!is_warning_error(2200));
+}
+
+#[test]
+fn test_order_message_routing() {
+    // Test OpenOrder with order ID at position 1
+    let message_str = "5\0123\0AAPL\0STK\0"; // OpenOrder with order_id=123
+    let message = ResponseMessage::from(message_str);
+    match determine_routing(&message) {
+        RoutingDecision::ByOrderId(id) => assert_eq!(id, 123),
+        routing => panic!("Expected ByOrderId routing, got {routing:?}"),
+    }
+
+    // Test CompletedOrdersEnd (no order ID)
+    let message_str = "102\01\0"; // CompletedOrdersEnd
+    let message = ResponseMessage::from(message_str);
+    match determine_routing(&message) {
+        RoutingDecision::ByOrderId(id) => assert_eq!(id, -1),
+        routing => panic!("Expected ByOrderId(-1) routing, got {routing:?}"),
+    }
+
+    // Test ExecutionData with order ID at position 2
+    let message_str = "11\01\0123\0456\0"; // ExecutionData with request_id=1, order_id=123
+    let message = ResponseMessage::from(message_str);
+    match determine_routing(&message) {
+        RoutingDecision::ByOrderId(id) => assert_eq!(id, 123),
+        routing => panic!("Expected ByOrderId routing, got {routing:?}"),
+    }
+
+    // Test CommissionsReport (no order ID but still an order message)
+    let message_str = "59\01\0exec123\0100.0\0USD\0"; // CommissionsReport
+    let message = ResponseMessage::from(message_str);
+    match determine_routing(&message) {
+        RoutingDecision::ByOrderId(id) => assert_eq!(id, -1),
+        routing => panic!("Expected ByOrderId(-1) routing, got {routing:?}"),
+    }
+}
